Move ProfileIter to codec::profile
